+++ tests/ote_cli/misc/ote_cli_test_common.py
@@ -15,8 +15,16 @@
 import os
 from subprocess import run  # nosec
 import logging
+from copy import deepcopy
 
 from ote_cli.utils.tests import get_some_vars, collect_env_vars
+from ote_cli.registry import Registry
+
+root = "/tmp/ote_cli/"
+ote_dir = os.path.dirname(
+    os.path.dirname(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
+)
+external_path = os.path.join(ote_dir, "external")
 
 default_train_args_paths = {
     "--train-ann-file": "data/airport/annotation_example_train.json",
@@ -35,6 +43,77 @@ wrong_paths = {
 }
 
 logger = logging.getLogger(__name__)
+
+
+def parser_templates():
+    params_values = []
+    params_ids = []
+    params_values_for_be = {}
+    params_ids_for_be = {}
+
+    for back_end_ in (
+            "DETECTION",
+            "CLASSIFICATION",
+            "ANOMALY_CLASSIFICATION",
+            "SEGMENTATION",
+            "ROTATED_DETECTION",
+            "INSTANCE_SEGMENTATION",
+    ):
+        cur_templates = Registry(external_path).filter(task_type=back_end_).templates
+        cur_templates_ids = [template.model_template_id for template in cur_templates]
+        params_values += [(back_end_, t) for t in cur_templates]
+        params_ids += [back_end_ + "," + cur_id for cur_id in cur_templates_ids]
+        params_values_for_be[back_end_] = deepcopy(cur_templates)
+        params_ids_for_be[back_end_] = deepcopy(cur_templates_ids)
+    return params_values, params_ids, params_values_for_be, params_ids_for_be
+
+
+def eval_args(_template_,
+              args_paths,
+              _ote_dir_,
+              _root_,
+              test_ann_file=True,
+              taf_path=None,
+              test_data_roots=True,
+              tdr_path=None,
+              l_weights=True,
+              lw_path=None,
+              save_performance=True,
+              sp_path=None,
+              additional=None):
+    _, twd, _ = get_some_vars(_template_, _root_)
+    ret_eval_args = [_template_.model_template_id]
+    if test_ann_file:
+        ret_eval_args.append("--test-ann-file")
+        if taf_path:
+            ret_eval_args.append(taf_path)
+        else:
+            ret_eval_args.append(f'{os.path.join(_ote_dir_, args_paths["--test-ann-files"])}')
+
+    if test_data_roots:
+        ret_eval_args.append("--test-data-roots")
+        if tdr_path:
+            ret_eval_args.append(tdr_path)
+        else:
+            ret_eval_args.append(f'{os.path.join(_ote_dir_, args_paths["--test-data-roots"])}')
+
+    if l_weights:
+        ret_eval_args.append("--load-weights")
+        if lw_path:
+            ret_eval_args.append(lw_path)
+        else:
+            ret_eval_args.append(f"{twd}/trained_{_template_.model_template_id}/weights.pth")
+
+    if save_performance:
+        ret_eval_args.append("--save-performance")
+        if sp_path:
+            ret_eval_args.append(sp_path)
+        else:
+            ret_eval_args.append(f"{twd}/trained_{_template_.model_template_id}/performance.json")
+
+    if additional:
+        ret_eval_args += [*additional]
+    return ret_eval_args
 
 
 def ote_common(template, root, tool, cmd_args):
