+++ tests/ote_cli/misc/test_ote_cli_export.py
@@ -44,8 +44,6 @@ external_path = os.path.join(ote_dir, "external")
 
 params_values = []
 params_ids = []
-params_values_for_be = {}
-params_ids_for_be = {}
 
 for back_end_ in (
     "DETECTION",
@@ -59,8 +57,6 @@ for back_end_ in (
     cur_templates_ids = [template.model_template_id for template in cur_templates]
     params_values += [(back_end_, t) for t in cur_templates]
     params_ids += [back_end_ + "," + cur_id for cur_id in cur_templates_ids]
-    params_values_for_be[back_end_] = deepcopy(cur_templates)
-    params_ids_for_be[back_end_] = deepcopy(cur_templates_ids)
 
 
 class TestExportCommon:
@@ -152,161 +148,3 @@ class TestExportCommon:
             assert (
                 error_string in ret["stderr"]
             ), f"Different error message {ret['stderr']}"
-
-
-class TestExportDetectionTemplateArguments:
-    @pytest.fixture()
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def create_venv_fx(self, template):
-        work_dir, template_work_dir, algo_backend_dir = get_some_vars(template, root)
-        create_venv(algo_backend_dir, work_dir)
-
-    @pytest.fixture()
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def get_pretrained_artifacts_fx(self, template, create_venv_fx):
-        get_pretrained_artifacts(template, root, ote_dir)
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def test_ote_export_pp_confidence_threshold_type(self, template, create_venv_fx):
-        _, template_work_dir, _ = get_some_vars(template, root)
-        error_string = "invalid float value"
-        command_args = [
-            template.model_template_id,
-            "--load-weights",
-            f"{template_work_dir}/trained_{template.model_template_id}/weights.pth",
-            "--save-model-to",
-            f"{template_work_dir}/exported_{template.model_template_id}",
-            "params",
-            "--postprocessing.confidence_threshold",
-            "String",
-        ]
-        ret = ote_common(template, root, "export", command_args)
-        assert ret["exit_code"] != 0, "Exit code must not be equal 0"
-        assert error_string in ret["stderr"], f"Different error message {ret['stderr']}"
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def test_ote_export_pp_confidence_threshold(
-        self, template, get_pretrained_artifacts_fx, create_venv_fx
-    ):
-        _, template_work_dir, _ = get_some_vars(template, root)
-        pre_trained_weights = (
-            f"{template_work_dir}/trained_{template.model_template_id}/weights.pth"
-        )
-        logger.debug(f"Pre-trained weights path: {pre_trained_weights}")
-        assert os.path.exists(
-            pre_trained_weights
-        ), f"Pre trained weights must be before test starts"
-        command_args = [
-            template.model_template_id,
-            "--load-weights",
-            pre_trained_weights,
-            "--save-model-to",
-            f"{template_work_dir}/exported_{template.model_template_id}",
-            "params",
-            "--postprocessing.confidence_threshold",
-            "0.5",
-        ]
-        ret = ote_common(template, root, "export", command_args)
-        assert ret["exit_code"] == 0, "Exit code must be equal 0"
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def test_ote_export_pp_confidence_threshold_oob(self, template, create_venv_fx):
-        _, template_work_dir, _ = get_some_vars(template, root)
-        error_string = "is out of bounds"
-        oob_values = ["-0.1", "1.1"]
-        for value in oob_values:
-            command_args = [
-                template.model_template_id,
-                "--load-weights",
-                f"{template_work_dir}/trained_{template.model_template_id}/weights.pth",
-                "--save-model-to",
-                f"{template_work_dir}/exported_{template.model_template_id}",
-                "params",
-                "--postprocessing.confidence_threshold",
-                value,
-            ]
-            ret = ote_common(template, root, "export", command_args)
-            assert ret["exit_code"] != 0, "Exit code must not be equal 0"
-            assert (
-                error_string in ret["stderr"]
-            ), f"Different error message {ret['stderr']}"
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def test_ote_export_pp_result_based_confidence_threshold_type(
-        self, template, create_venv_fx
-    ):
-        _, template_work_dir, _ = get_some_vars(template, root)
-        error_string = "Boolean value expected"
-        command_args = [
-            template.model_template_id,
-            "--load-weights",
-            f"{template_work_dir}/trained_{template.model_template_id}/weights.pth",
-            "--save-model-to",
-            f"{template_work_dir}/exported_{template.model_template_id}",
-            "params",
-            "--postprocessing.result_based_confidence_threshold",
-            "NonBoolean",
-        ]
-        ret = ote_common(template, root, "export", command_args)
-        assert ret["exit_code"] != 0, "Exit code must not be equal 0"
-        assert error_string in ret["stderr"], f"Different error message {ret['stderr']}"
-
-    @e2e_pytest_component
-    @pytest.mark.parametrize(
-        "template",
-        params_values_for_be["DETECTION"],
-        ids=params_ids_for_be["DETECTION"],
-    )
-    def test_ote_export_pp_result_based_confidence_threshold(
-        self, template, get_pretrained_artifacts_fx, create_venv_fx
-    ):
-        _, template_work_dir, _ = get_some_vars(template, root)
-        pre_trained_weights = (
-            f"{template_work_dir}/trained_{template.model_template_id}/weights.pth"
-        )
-        logger.debug(f"Pre-trained weights path: {pre_trained_weights}")
-        assert os.path.exists(
-            pre_trained_weights
-        ), f"Pre trained weights must be before test starts"
-        command_args = [
-            template.model_template_id,
-            "--load-weights",
-            pre_trained_weights,
-            "--save-model-to",
-            f"{template_work_dir}/exported_{template.model_template_id}",
-            "params",
-            "--postprocessing.result_based_confidence_threshold",
-            "False",
-        ]
-        ret = ote_common(template, root, "export", command_args)
-        assert ret["exit_code"] == 0, "Exit code must be equal 0"
