+++ tests/ote_cli/misc/test_ote_cli_optimize.py
@@ -397,7 +397,7 @@ class TestOptimizeDetectionTemplateArguments:
     )
     def create_venv_fx(self, template):
         work_dir, template_work_dir, algo_backend_dir = get_some_vars(template, root)
-        create_venv(algo_backend_dir, work_dir, template_work_dir)
+        create_venv(algo_backend_dir, work_dir)
 
     @pytest.fixture()
     @e2e_pytest_component
